Сделал парсинг облигации по ссылке Smart-Lab.

+++ BondYieldCalculator/BondYieldCalculator.Entities/CommonBondInfo.cs
@@ -6,6 +6,6 @@
 
         public decimal CurrentPrice { get; set; }
 
-        public double Maturity { get; set; }
+        public decimal Maturity { get; set; }
     }
 }
